requirements.txt
opencv-python~=4.8.1.78
numpy~=1.26.1
ultralytics~=8.1.47
python-telegram-bot
roboflow
wandb